+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2018-10-01
+### Added
+- Support for `--from` and `--till` options
+- Support for tag-names as `latest-of-major`, `latest-of-major-x` (where
+  x is the delta minus)
+
 ## [1.0.1] - 2018-05-25
 ### Fixed
 - Errors for missing files